requirements.txt
fastapi==0.110.2
uvicorn==0.30.1
transformers==4.41.1
torch>=2.1.0
keybert==0.8.3
sentence-transformers==2.6.1
scikit-learn==1.5.0
pydantic==2.7.1
numpy==1.26.4
psycopg2-binary==2.9.10
streamlit==1.46.1
pandas==2.3.0
plotly==6.2.0
